server/requirements.txt
Flask
pandas
flask-cors
google-generativeai
